feat: add cjsToEsm.mjs solution
